requirements.txt
pandas>=2.0
numpy>=1.24
aiohttp>=3.9
httpx>=0.27
sqlalchemy>=2.0
aiosqlite>=0.19
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
python-multipart>=0.0.9
plotly>=5.20
openpyxl>=3.1
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
